mutablelimiter: Fix flakey test (#4161)
If a machine has high load there is a chance that TestLimiter would
fail. "ctx3" has a timeout of 250ms, which is low enough that the call to
SetLimit won't cancel contexts in time. So we increase the timeout drastically
to reduce the chance of the flakey failure.

Additionally we use a better pattern for waiting on "ctx1" to be cancelled.