properly handle instances that have empty settings (#4157)
In older (2.x) versions of Sourcegraph it was possible to save an empty string
as your user/org/global settings. This lead to various issues including just
completely breaking the settings page: https://github.com/sourcegraph/sourcegraph/issues/1645

We later fixed that issue and made it such that saving empty settings is
forbidden, but it doesn't appear we ever corrected such invalid settings. In
one customer environment we've encountered a user who is reporting their
settings are empty so we should handle this.

This is a quick fix which is cheap for us to maintain, so I opted for this
approach for now. The more correct (but also more complex) approach would be to
write a migration which transforms any empty settings into non-empty ones. I
don't think it is worth the time it would take right now.