pkg/store: corrupted tars from git are ignored
In some cases gitserver would return a corrupted archive. It seems this would
be an ephemeral issue. This would cause searches to fail. However, it should
be treated as a temporary error. This change ensures the failure is marked as
temporary, leading to the graphql layer to not fail the full search.