web: Remove references to enabled in graphql queries
For fetching the field, it is deprecated and always true. We have already
removed all reads of the field in the codebase.

For specifying enabled/disabled in repo list filters, we have updated the
backend to default enabled to true. Additionally we have removed the option
for the user to request enabled=false or disabled=true in the UI.