all: gofumpt
gofumpt is a stricter format than gofmt, while being backwards
compatible. That is, gofumpt is happy with a subset of the formats that gofmt
is happy with. See https://github.com/mvdan/gofumpt

  gofumpt -w -s .

We won't enforce gofumpt output since that requires all developers of
sourcegraph to switch to it. Instead this commit is a once off (and may happen
again in the future).