repo-updater: Deterministic merge and prevent name conflicts from sources (#4283)
Depending on the order we iterate through sourced, the final "Repo" could
change. Previously we didn't worry about this since we assumed eventual
consistency from sources (on external ID). However, we are removing this
assumption in the interest of more robustness.

Additionally we can have case-insensitive conflicts on name even after
merging all repos with the same external ID. So we deterministically pick a
winner when such a conflict occurs.