gitserver: Improve perf by not specifying git protocol version
The blog post mentioned in the code that's being removed here got an
update in December 2018 that specifies:

    Previously the example above used the more rarely used "git ls-remote"
    command. A recent change to improve the correctness of that command's
    output causes it to request more references.

This change causes `git ls-remote` with `protocol.version=2` to be
consistently _slower_ than the variant without a protocol specified.

In order to improve performance, we remove the code and use the older
protocol for ls-remote.