web: Show a 'Clone now' button in repos list if not cloned yet
This fixes #4232 by making it clearer to the user that they can speed up
the cloning process by clicking on the "Clone now" button.

We (@keegancsmith and I) discussed other solutions (showing the position
in the clone schedule, rewording "Not yet cloned" label, etc.) but
ultimately came to the conclusion that "Not yet cloned" is correct and
that showing any position in a cloning-schedule would be unreliable.

"Clone now" on the other hand makes the tooltip text of the "Not yet
cloned" label actionable as a button and gives the user something to do.