Refactor trackViews() to reliably remove views
Fixes #4064

Refactors `trackViews()` to stop relying on a `ViewResolver`'s selector to detect
removed views. We now inspect `removedNodes` to see if any of the removed nodes
is a known view, or contains one or more known views.

With this refactor, the `viewStates` logic is now encapsulated within `trackViews()`,
which simply emits every added views as a `ViewWithSubscriptions`, and takes care of calling
`view.subscriptions.unsubscribe()` when the view is removed from the page. This prevents
the consumer from having to reimplement the `viewStates` pattern.