diskcache: fsync cache item and cache directory
We have seen cases where corrupt zip files exist on disk. We fixed that by
evicting bad archives. However, this was likely due to OS caches not being
flushed. We now flush both the file and the directory.

For context on why we call fsync see https://lwn.net/Articles/457667/