pkg/conf: add DISABLE_CONF_DEADLOCK_DETECTOR=true for debugging
This allows you to disable the pkg/conf deadlock detector in dev mode which
triggers anytime the frontend does not start after a reasonable amount of time.

Primarily this is useful for debugging e.g. DB migrations that take far too long,
or other causes of the frontend not starting in a timely manner in dev mode.

Fixes #4380