pkg/trace: include GraphQL ?Type parameter in route name
Today our HTTP metrics are pretty nice, but when it comes to any GraphQL requests
they are all grouped under the `graphql` route which makes them mostly useless.

For example, you can see HTTP latency is quite bad or that there are lots of error
codes on GraphQL requests -- but it is impossible to know if those are search
requests or some super-slow request made on e.g. the admin-only repository list
page.

I chose to include this in the route name (`graphql: Search`) instead of adding a
different field because it is nice to not have to deal with the separate field in
dashboard queries and because the route field would be mostly useless on GraphQL
requests otherwise.

Fixes #4246