Faster Webpack builds (#4448)
* print Webpack "Using mode ..." to stderr

This prevents it from polluting `webpack --json` output (which is used for profiling Webpack perf and bundle size).

* remove unneeded RxJS Webpack alias

This added ~4s to the initial Webpack build time and ~400msec to the incremental rebuild time.

* remove unnecessary/ineffective warnings filter

This does not work and is no longer necessary because the new way that monaco-editor is packaged removes this warning.

* dsable Webpack optimizations in dev

This reduces incremental rebuild time by ~600msec.

* do not use mini-css-extract-plugin in development

In production mode, we still use https://github.com/webpack-contrib/mini-css-extract-plugin to extract CSS to a separate CSS file for better caching and faster client rendering.

We used to *also* use mini-css-extract-plugin in development, but this is not advised. From https://github.com/webpack-contrib/mini-css-extract-plugin#advanced-configuration-example:

> This plugin should be used only on production builds ...

In development, we do NOT want to use mini-css-extract-plugin because it slows down Webpack (initial builds by ~5s and incremental reloads by ~2s).

There is a slight impact on debuggability in browser devtools. In development, CSS styles will now appear to come from inline `<style></style>` tags, not from a `.css` file with sourcemaps. The benefit of *much* faster build times (for both TypeScript and SCSS) outweighs this downside.