Browser extension: use the raw repo name when getting the definition URL
Fixes #4368

Location URIs emitted by registered location providers may contain the
repo name as affected by `repositoryPathPattern`, rather than the raw repo
name as inferred from the code host's DOM / external service.

This was an issue when building jump-to-definition URLs, because the browser
extension code would assume that the `repoName` passed to `urlToFile()` contained
the code host's hostname, leading to #4368.

This is fixed by ensuring that the raw repo name is resolved and passed down
to `urlToFile()`, then properly handled to ensure that we only build valid
jump-to-definition URLs.