flatten admin area in DiscussionsComment (#4462)
Previously, there was a rounded area with a lock icon and admin actions (clear reports, and delete the comment forever). By far the most common state was for there to be only 1 admin action because "clear reports" only appears when there are reports (ie complaints) on a comment. This was confusing because the lock icon seemed just as clickable as the icon next to it.

Now, the action icons are all shown adjacent to each other. This is sufficient and removes code and visual complexity.