Display a dismissable alert in the hover
This improves discoverability of the `codeHost.useNativeTooltips` setting by displaying a dismissable, one-time alert in the hover content informing the user that we're hiding the code host's native tooltips.

In order to make this reasonably clean (and not specific to this alert), I added the possibility for the `HoverOverlay` to display arbitrary dismissable alerts. Currently, only the browser extension implements it, keeping track of dismissed alerts through `storage.sync`.