use eslint rules for React hooks (#4486)
TSLint is being deprecated in favor of ESLint (https://github.com/palantir/tslint/issues/4534). We are not yet ready to fully migrate to ESLint because we rely on some TSLint-specific rules, but there is one set of particularly valuable ESLint-only rules: https://www.npmjs.com/package/eslint-plugin-react-hooks.

This introduces ESLint to our codebase and uses it only for those rules, to avoid duplication with TSLint.