remove duplicate PopoverButton impl
There were 2 impls of PopoverButton: the original in `./web` and the new(er) one in `./shared`. This removes the original (decayed) impl and uses the shared impl for all popovers in the web app. The only loss of functionality is the unadvertised and little-used `r` and `v` hotkeys for opening the repo header repo list popover and rev list popover, respectively. This is acceptable.