upgrade+unvendor reactstrap, remove PopoverButton (#4545)
* use upstream reactstrap

The bug that required us to use a vendored version was fixed (see #1553).

fix #1553

* remove duplicate PopoverButton impl

There were 2 impls of PopoverButton: the original in `./web` and the new(er) one in `./shared`. This removes the original (decayed) impl and uses the shared impl for all popovers in the web app. The only loss of functionality is the unadvertised and little-used `r` and `v` hotkeys for opening the repo header repo list popover and rev list popover, respectively. This is acceptable.

* remove PopoverButton, use Bootstrap's Popover instead

Our custom PopoverButton class added unnecessary complexity, and this commit removes it.

Due to https://github.com/reactstrap/reactstrap/issues/1484 (introduced in the Reactstrap upgrade), this required a bit of hacking.

Confirmed working via the browser extension on GitHub and GitLab.