multiline search query field (textarea not input)
This makes the global and per-repo search query field support multiple lines (in the UI only; this is not a backend change). When the query is too long to fit on one line, the input automatically expands down one line to fit it. This is similar to how the message field in Slack, etc., works. It is nicer than requiring the user to scroll left and right to see their full query.

Users can also force a new line by pressing Shift+Enter.

This is NOT about supporting queries that match across multiple lines.