Support toggling display of a code host's native tooltips (#4495)
Fixes #4479

This avoids conflicts between GitHub's new jump to definition tooltips, and the Sourcegraph tooltips.

It introduces a new setting, `codeHost.useNativeTooltips`, and uses it to determine whether to show the code host's hover tooltips, or our own. By default, the browser extension will hide the code host's native tooltips, and display our own. If the user sets `codeHost.useNativeTooltips` to `true` in his settings, only the code host's hovers will be displayed.

To improve discoverability of the `codeHost.useNativeTooltips`, we display a dismissable, one-time alert in the hover content informing the user that we're hiding the code host's native tooltips.

In order to make this reasonably clean (and not specific to this alert), I added the possibility for the `HoverOverlay` to display arbitrary dismissable alerts. Currently, only the browser extension implements it, keeping track of dismissed alerts through `storage.sync`.