dev: Fix `notest` docker build branches
The idea behind the docker-images-patch-notest/<IMAGE> branches was that
you could push to them to build a Docker image on CI without running
through the test suite (hence the notest in the name).

All this change here does is move the existing code around so that the
check for the branch name results in the early exit that we want.

To test I pushed this branch here to the a notest-branch:

$ git push origin
docker/fix-notest-branches:docker-images-patch-notest/frontend --force

The resulting build only generates a pipeline and runs a Docker image
build.