Declare the GraphQL search pagination API
Following up from https://github.com/sourcegraph/sourcegraph/pull/4230 this is
the next logical step in the "cursor-based result-level pagination API"
mentioned at the end of the Google doc.

You might be asking: "Where is the cursor? I only see `offset` and `limit`."
and that is a good question. I believe we do not actually need an explicit
cursor as part of the API either because in the backend we can already make use
of an implicit cursor defined as `(query, userID)`. That is, if the following
hold true:

1. Search results have a well-defined global order.
2. We do not care to immediately solve the problem of "what happens if new
   search results are introduced _as we are paginating_?" (we will have some
   well-defined flexibility here)
3. The union `(query, userID)` remains an accurate depiction of a total result
   set (it is). Note that we include the `userID` because search results are
   tailored to the user (e.g. in the case user A has access to repositories
   user B does not then we would not want to share result sets.

There is room for a possible future optimization where we could share results
between users executing the same query and with access to the same result sets,
but that is true today as well without this change in place.