(final): A cursor-based API
This was the original concluding direction of my research document, and is the API
I am proposing. You can think of this as streaming over GraphQL, but it is
different in a few important ways I will outline below.

Like streaming, when you begin performing a search query you must start at
result zero. While we would leave this limitation in place, it is important to
note it is being enforced in the backend and not a restriction of the API
itself. We _could_ offer an offset like `offset: Int!` or `SearchOffset(repositories: Int!, results: Int!)`
as in my prior two commits here -- but it comes with the same limitations: we
would be locking ourselves into a bad API.

Unlike streaming, a cursor is not an _active connection_:

- A cursor is just an opaque offset structure (an arbitrary ID string).
- Initially in our backend it would be defined as `(repositoryOffset, resultOffset, userID)`.
- You can come back with the same cursor hours or even days later, and it will
  still work (there are some edges to define here, of course, like what happens
  when new results are introduced or repositories are deleted).
- The `userID` being a part of the cursor is important because the `repositoryOffset`
  is based on the repositories that the user has access to via ACL enforcement.
- While it is not a connection, our backend can make smart choices like
  pre-fetching results ahead of the user so that their next request is fast
  (unless they've been gone for a long time, in which case we could restart at
  the repository offset denoted in the cursor).
- A cursor could be shared between users. This is important for multiple reasons
  associated with sharing:
  a. Alice shares a link to result 9,000 with Bob, Bob doesn't want to first fetch 9,000
    results.
  b. A tool built against Sourcegraph's API find an interesting search results
     and notifies a user (for example, Saved Searches would use this API) ->
     that user will not want to wait for all prior search results to be loaded.