(exploratory): An API based on SearchOffset(repo, result)
My prior approach from last night was incorrect (I just forgot the details I
had already learned): We cannot use `(offset, limit)` as our search pagination
API for result-level pagination because we do not fundamentally know how many
results exist in a repository until we've completed the search in that repository
already (at which point, why paginate?)

This is an API that would work to solve the issue. The idea is that you are
specifying both a number of repositories and number of results (in the last
repository) to offset by, which allows us to avoid the work needed searching
in a repository just to know how many results it contains.

The major downside of this approach is that we become locked-in to an API that
does repository-level ordering. For example, when we eventually add smart
ordering of results and want to rank three results ("best", "okay", and "bad")
across two repositories, we would be limited to doing so within a single
repository like so:

1. `repoA` -> `best`
2. `repoA` -> `bad`
3. `repoB` -> `okay`

i.e. result ordering can be done inside repositories, but not across repositories.

Since this would (obviously) be very bad, it is not the API I am proposing. See
my next commit.