repo-updater: Fix not-cloned count by comparing lowercase repo names
`gitserver`'s `list?cloned` endpoint only returns lowercase repository
names. In order for the computation of the not-cloned count in the
status indicator to work properly, we need to lowercase the repository
names we load from the database before comparing them to gitservers.

Discovered this by accident when cloning the complete
`github.com/sourcegraph` organization and seeing the not-cloned count
not going down after a certain point.