Enable clone status indicator by default
The original reason for putting the status indicator was that on large
instances it was always in an active state, indicating that repositories
are being cloned, when in fact they were only enqueued for updates. See
PR #4497 for more details on this.

With the improvements made in the following PRs, I think it's safe to
enable the status indicator by default now:

- #4591 - status-indicator: Show actual cloning state
- #4810 - perf: Improve performance of computing number of not-cloned repos
- #4859 - Fix not-cloned count by comparing lowercase repo names

The number is now accurate and it's reasonably fast to compute (3.3s on
an instance with 33k repositories, calculated every 30s).

In the next release we can remove the feature flag altogether.