Allow creating and updating (user|org) names with dots
In #4690, we made it possible to create usernames from SSO signups that include dots.

But it was still not possible to change or manually create an account with a username that
includes a dot in the UI. This PR fixes that.

For consistency, we allow dots in org names too.

Fixes #4909, #4920