inject routes into OrgArea in same way as for UserArea
There was duplicate code for the user and org nav items and routes: the user nav items and routes were extracted into navitems.ts and routes.tsx, and the org ones weren't.

This refactors the org area code to be consistent with how this is done for the user area, which eliminates some duplication whenever anything (such as settings or saved searches) needs to be present in both the user and org areas.

This also fixes 2 minor issues: the "Saved searches" tab was visible for viewers who don't have admin perms on the
org (it is not visible for the user, which is correct); and the order of the "saved searches" tab and "settings" tab is inverted between users and orgs.