Add e2e tests for symbol table to ensure symbols change with file
This implements E2E tests for the following 3 items in the release
testing grid:

* `7.4.1 Verify the symbol sidebar works in each of the following languages: Go`
* `7.4.2 Verify the symbol sidebar works in each of the following languages: TypeScript`
* `7.4.3 Verify the symbol sidebar works in each of the following languages: Python`

The added cost are the two new repos, `appdash` and
`sourcegraph-typescript`. But those are necessary to have so we can test
that the symbol bar works for Go, TypeScript and Python.

Regarding the tests themselves: I'm not sure how welcome "table-driven
tests" such as these are in the TypeScript world, but since we want to
test the same thing for multiple files in different languages I think
this is the most readable and extendable version.