Check for lineStart from lastStart rather than lastLineStart
If you have a long line with lots of matches, the overall algorithm for
finding the lineStart degenerates into O(n^2) since we keep searching for a
newline over content we already know doesn't have new lines. IE
lastLineStartIndex never changes. But start always changes. I can write a
proof with invariants on why this works and is better! However, I think once
we have cleaned up this PR it would be better to annotate the code with
invariants (like how the Sort implementation in the stdlib does).