all: go test skip on missing tools
A good property we used to have (and should keep) is that `go test -short ./...` just
works, and `go test ./...` works if you have our important services running
(redis/postgres). This PR fixes up some cases where this could fail.

In particular if you don't have comby on your path the test skips (we use the
same pattern for other tool dependencies). We also use a custom build of
universal-ctags for symbols, so if you have the normal version on your path we
don't fail tests due to it failing to start.

In both cases we do fail if we are on CI, to ensure we don't regress.