authz: Non-blocking Bitbucket Server permissions updates (#5180)
* authz: Non-blocking Bitbucket Server permissions updates

This commit changes the Bitbucket Server permissions store to perform
non-blocking permissions updates. When there are existing expired
permissions, those are used immediatelly, and a background update is
started. Otherwise, an error is returned to the user indicating that
their permissions are stale.

This commit also changes the locking mechanism used to control
concurrency of background updates so that it doesn't block. This
prevents exhaustion of the Postgres connection pool.

A follow-up PR will introduce the concept of hard and soft TTLs as
described in #4812.

* fixup! Remove race in tests

* fixup! Document thread safety

* Update enterprise/cmd/frontend/internal/authz/bitbucketserver/store.go

Co-Authored-By: Keegan Carruthers-Smith <keegan.csmith@gmail.com>

* Update enterprise/cmd/frontend/internal/authz/bitbucketserver/store.go

* fixup! Create a named type for the update closure

* fixup! Improve naming

* fixup! Explictly set IDs to nil

* fixup! Explict advisory lock namespace

* fixup! Ensure that lock is called within a transaction

* fixup! Do not log errLockNotAvailable

* fixup! Do not update cache in blocking code path

This makes the blocking and non blocking paths more similar

* fixup! Use pg_try_advisory_xact_lock

* fixup! Cast uint32 to int32

* fixup! cast lockNamespace only once