use GraphQL DateTime scalar type for dates
Previously, GraphQL datetime values were `String` and their format was only specified in (inconsistent) docstrings. Now, all datetime values are of the GraphQL scalar type `DateTime`, which is documented in one place as being RFC3339.

There is no change to API consumers (the API responses remain exactly the same for all queries); this is purely an API documentation and organization improvement.