emit React hot looader wrappers for Area components
For React hot loading (which is when, after saving a `.ts`/`.tsx` file in your editor, your browser loads new React components and inserts them into the page in place of the old components), each code splitting entrypoint must have the wrapper applied. The most common level at which we perform code splitting is the "Area" (ie user area, settings area, etc.). This change makes it so that more changes in local dev can be hot-reloaded.