move registry side-effects import to shared
This package *is* indeed imported and required by both the OSS and enterprise builds, so it should be in shared. In practice, this does not make any difference in the final binary because the enterprise build already imported this package. But moving it to the correct place (shared) is still good because it documents the intent (for it to be shared, not for it to be OSS-only).